+++ lib/astarte_rpc/protocol/proto/housekeeping/get_realm_reply.pb.ex
@@ -1,9 +1,7 @@
 defmodule Astarte.RPC.Protocol.Housekeeping.GetRealmReply.DatacenterReplicationFactorsEntry do
   @moduledoc false
 
-  use Protobuf, map: true, protoc_gen_elixir_version: "0.11.0", syntax: :proto3
-
-  def fully_qualified_name, do: "GetRealmReply.DatacenterReplicationFactorsEntry"
+  use Protobuf, map: true, syntax: :proto3, protoc_gen_elixir_version: "0.12.0"
 
   field :key, 1, type: :string
   field :value, 2, type: :int32
@@ -12,9 +10,7 @@ end
 defmodule Astarte.RPC.Protocol.Housekeeping.GetRealmReply do
   @moduledoc false
 
-  use Protobuf, protoc_gen_elixir_version: "0.11.0", syntax: :proto3
-
-  def fully_qualified_name, do: "GetRealmReply"
+  use Protobuf, syntax: :proto3, protoc_gen_elixir_version: "0.12.0"
 
   field :realm_name, 1, proto3_optional: true, type: :string, json_name: "realmName"
   field :jwt_public_key_pem, 2, proto3_optional: true, type: :string, json_name: "jwtPublicKeyPem"
@@ -35,4 +31,8 @@ defmodule Astarte.RPC.Protocol.Housekeeping.GetRealmReply do
     proto3_optional: true,
     type: :int64,
     json_name: "deviceRegistrationLimit"
+
+  field :datastream_maximum_storage_retention, 7,
+    type: :int64,
+    json_name: "datastreamMaximumStorageRetention"
 end